fix #7829 (test for 7810 on 32-bit)

+++ test/core.jl
@@ -1812,4 +1812,4 @@ type Foo7810{T<:AbstractVector}
     v::T
 end
 bar7810() = [Foo7810([(a,b) for a in 1:2]) for b in 3:4]
-@test Base.return_types(bar7810,())[1] == Array{Foo7810{Array{(Int64,Int64),1}},1}
+@test Base.return_types(bar7810,())[1] == Array{Foo7810{Array{(Int,Int),1}},1}
